refactor(tooling): split task helper modules
